requirements.txt
pandas==2.3.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
numpy==2.2.6
matplotlib==3.10.3
seaborn==0.13.2
plotly==6.2.0
ipywidgets==8.1.7